+++ examples/chat.nim
@@ -7,7 +7,10 @@ import pkg/bard
 
 dotenv.load()
 
-let ai = waitFor newBardAi getEnv "bard_cookies"
+let ai = waitFor newBardAi({
+  "__Secure-1PSID": getEnv("bard_psid"),
+  "__Secure-1PSIDTS": getEnv("bard_psidts")
+})
 
 var chat = ai.newBardAiChat